pyproject.toml
[project]
name = "spellchecker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=25.1.0",
    "python-levenshtein>=0.26.1",
]
